pages component 切り分け
